ref(agno): Make agno use new integrations API
