+++ py/src/braintrust/integrations/__init__.py
@@ -1,5 +1,6 @@
 from .adk import ADKIntegration
+from .agno import AgnoIntegration
 from .anthropic import AnthropicIntegration
 
 
-__all__ = ["ADKIntegration", "AnthropicIntegration"]
+__all__ = ["ADKIntegration", "AgnoIntegration", "AnthropicIntegration"]
