+++ py/src/braintrust/integrations/agno/cassettes/test_auto_agno.yaml
@@ -1,0 +1,147 @@
+interactions:
+- request:
+    body: '{"messages":[{"role":"developer","content":"You are a helpful assistant.
+      Be brief."},{"role":"user","content":"Say hi"}],"model":"gpt-4o-mini"}'
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate, zstd
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '143'
+      Content-Type:
+      - application/json
+      Host:
+      - api.openai.com
+      User-Agent:
+      - PatchedOpenAI/Python 2.15.0
+      X-Stainless-Arch:
+      - arm64
+      X-Stainless-Async:
+      - 'false'
+      X-Stainless-Lang:
+      - python
+      X-Stainless-OS:
+      - MacOS
+      X-Stainless-Package-Version:
+      - 2.15.0
+      X-Stainless-Raw-Response:
+      - 'true'
+      X-Stainless-Runtime:
+      - CPython
+      X-Stainless-Runtime-Version:
+      - 3.13.3
+      x-stainless-read-timeout:
+      - '600'
+      x-stainless-retry-count:
+      - '0'
+    method: POST
+    uri: https://api.openai.com/v1/chat/completions
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAwAAAP//jFJBbtswELzrFVuerUKWjdjyJZcWSBG0QC9F0SIQaHIls6G4LLlqawT+
+        e0HJsZQ0BXLRYWdnNDPchwxAGC12INRBsuq8zd8Vunj/8RPef9k3n5tN87W3+ue3Zh1u+TaKRWLQ
+        /gcqfmS9VdR5i2zIjbAKKBmT6nJzVRXrsqy2A9CRRptored8TXlnnMnLolznxSZfbs/sAxmFUezg
+        ewYA8DB8k0+n8Y/YQbF4nHQYo2xR7C5LACKQTRMhYzSRpWOxmEBFjtEN1m/MG7ih36Ckgw8wbsOR
+        emDS8ng9ZwVs+iiTc9dbOwOkc8QyJR/83p2R08WhpdYH2sdnVNEYZ+KhDigjueQmMnkxoKcM4G5o
+        on8STvhAneea6R6H35XlKCem/iewOmNMLO00Xi0XL4jVGlkaG2dFCiXVAfXEnFqXvTY0A7JZ5H+9
+        vKQ9xjaufY38BCiFnlHXPqA26mneaS1gOs7/rV0qHgyLiOGXUVizwZCeQWMjezuejIjHyNjVjXEt
+        Bh/MeDeNr8tqtSpkdbXdiuyU/QUAAP//AwBBs78WRQMAAA==
+    headers:
+      CF-RAY:
+      - 9c1afcdd3f36b231-SJC
+      Connection:
+      - keep-alive
+      Content-Encoding:
+      - gzip
+      Content-Type:
+      - application/json
+      Date:
+      - Thu, 22 Jan 2026 00:38:19 GMT
+      Server:
+      - cloudflare
+      Set-Cookie:
+      - __cf_bm=Jw95ZRGfTr6qO8YVvMCpB1aMAiti.HWb9WM0o.EAG4M-1769042299-1.0.1.1-F0ol4YtLGC1.t2DHb1Hj435gvyQ_nGNudwYUErS.pg4aWKbU4O68f4wJthw2GUCv2BYU7cC4ZcIA0B6TvaUN7VYsBM5OS7Ccc46cnb7zQ9Y;
+        path=/; expires=Thu, 22-Jan-26 01:08:19 GMT; domain=.api.openai.com; HttpOnly;
+        Secure; SameSite=None
+      - _cfuvid=gxrFvllhyUbQeecWVXMHkFhdg_IAJ7CO467JJDSyVA8-1769042299331-0.0.1.1-604800000;
+        path=/; domain=.api.openai.com; HttpOnly; Secure; SameSite=None
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      Transfer-Encoding:
+      - chunked
+      X-Content-Type-Options:
+      - nosniff
+      access-control-expose-headers:
+      - X-Request-ID
+      alt-svc:
+      - h3=":443"; ma=86400
+      cf-cache-status:
+      - DYNAMIC
+      openai-organization:
+      - braintrust-data
+      openai-processing-ms:
+      - '438'
+      openai-project:
+      - proj_vsCSXafhhByzWOThMrJcZiw9
+      openai-version:
+      - '2020-10-01'
+      x-envoy-upstream-service-time:
+      - '490'
+      x-openai-proxy-wasm:
+      - v0.1
+      x-ratelimit-limit-requests:
+      - '30000'
+      x-ratelimit-limit-tokens:
+      - '150000000'
+      x-ratelimit-remaining-requests:
+      - '29999'
+      x-ratelimit-remaining-tokens:
+      - '149999985'
+      x-ratelimit-reset-requests:
+      - 2ms
+      x-ratelimit-reset-tokens:
+      - 0s
+      x-request-id:
+      - req_c68836a69b1549819fb6a5eecfd10be7
+    status:
+      code: 200
+      message: OK
+- request:
+    body: '{"session_id":"3ed01154-18cc-4648-b766-73f60e3e08c2","run_id":"4ebf7a0f-31fa-4a69-9500-f3f3f21d350d","data":{"agent_id":"test-agent","db_type":null,"model_provider":"OpenAI","model_name":"OpenAIChat","model_id":"gpt-4o-mini","parser_model":null,"output_model":null,"has_tools":true,"has_memory":false,"has_learnings":false,"has_culture":false,"has_reasoning":false,"has_knowledge":false,"has_input_schema":false,"has_output_schema":false,"has_team":false},"sdk_version":"2.4.1","type":"agent"}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate, zstd
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '493'
+      Content-Type:
+      - application/json
+      Host:
+      - os-api.agno.com
+      user-agent:
+      - agno/2.4.1
+    method: POST
+    uri: https://os-api.agno.com/telemetry/runs
+  response:
+    body:
+      string: '{"message":"Run creation acknowledged: 4ebf7a0f-31fa-4a69-9500-f3f3f21d350d","status":"success"}'
+    headers:
+      content-length:
+      - '96'
+      content-type:
+      - application/json
+      date:
+      - Thu, 22 Jan 2026 00:38:19 GMT
+      server:
+      - uvicorn
+    status:
+      code: 201
+      message: null
+version: 1
